refactor : riotDto 리팩토링 및 geminiSerive 수정
