+++ backend/src/main/java/org/mtvs/backend/riot/dto/PositionDto.java
@@ -1,0 +1,38 @@
+package org.mtvs.backend.riot.dto;
+
+public class PositionDto {
+    private int x;
+    private int y;
+
+    public PositionDto() {
+    }
+
+    public PositionDto(int x, int y) {
+        this.x = x;
+        this.y = y;
+    }
+
+    public int getX() {
+        return x;
+    }
+
+    public void setX(int x) {
+        this.x = x;
+    }
+
+    public int getY() {
+        return y;
+    }
+
+    public void setY(int y) {
+        this.y = y;
+    }
+
+    @Override
+    public String toString() {
+        return "PositionDto{" +
+                "x=" + x +
+                ", y=" + y +
+                '}';
+    }
+}
